List critical dyads

Critical dyads are edges on a specified level which break motifs by being removed from the network.

critical_dyads(net, motif, lvl_attr = c("sesType"), level = -1)

Arguments

net
network object

motif
motif identifier

lvl_attr
character vector specifying the attribute name where level information is stored in net

level
level of the dyads which shall be considered, or -1 if the level shall be determined automatically.

Value

data frame with three columns, listing edges and their contribution to motifs described by the motif identifier in descending order

Details

The level parameter determines on which level of the network critical dyads are analysed. Per default, when level = -1, the first level in the motif which provides exactly two nodes is selected. Use this parameter to specify a level manually. The procedure for determining the level is the same as for the Actor's Choice Model, cf. vignette.

Note that this only works for undirected graphs. Regardless of whether the input graph is directed it is treated as undirected graph.

Examples

# \dontrun{
critical_dyads(ml_net, motif = "1,2[I.C]")

#>     vertex0 vertex1 contribution
#> 1   actor16 actor24            4
#> 2    actor6  actor7            3
#> 3    actor6 actor54            3
#> 4    actor7 actor54            3
#> 5   actor10 actor18            3
#> 6   actor10 actor24            3
#> 7   actor16 actor37            3
#> 8   actor16 actor58            3
#> 9   actor24 actor37            3
#> 10  actor24 actor52            3
#> 11  actor25 actor58            3
#> 12  actor31 actor44            3
#> 13  actor52 actor55            3
#> 14   actor2 actor16            2
#> 15   actor2 actor24            2
#> 16   actor2 actor51            2
#> 17   actor3 actor27            2
#> 18   actor3 actor56            2
#> 19   actor4 actor56            2
#> 20   actor5 actor44            2
#> 21   actor7 actor27            2
#> 22   actor8 actor10            2
#> 23   actor9 actor10            2
#> 24   actor9 actor44            2
#> 25  actor10 actor14            2
#> 26  actor10 actor16            2
#> 27  actor10 actor37            2
#> 28  actor10 actor44            2
#> 29  actor10 actor51            2
#> 30  actor10 actor56            2
#> 31  actor14 actor56            2
#> 32  actor15 actor27            2
#> 33  actor15 actor41            2
#> 34  actor15 actor54            2
#> 35  actor16 actor51            2
#> 36  actor18 actor30            2
#> 37  actor18 actor56            2
#> 38  actor24 actor51            2
#> 39  actor25 actor50            2
#> 40  actor25 actor52            2
#> 41  actor25 actor55            2
#> 42  actor27 actor54            2
#> 43  actor27 actor56            2
#> 44  actor50 actor58            2
#> 45  actor55 actor58            2
#> 46   actor1  actor9            1
#> 47   actor1 actor25            1
#> 48   actor1 actor36            1
#> 49   actor1 actor44            1
#> 50   actor1 actor46            1
#> 51   actor1 actor58            1
#> 52   actor2 actor56            1
#> 53   actor3 actor41            1
#> 54   actor4  actor9            1
#> 55   actor4 actor13            1
#> 56   actor4 actor28            1
#> 57   actor4 actor44            1
#> 58   actor5  actor9            1
#> 59   actor7 actor41            1
#> 60   actor9 actor13            1
#> 61   actor9 actor14            1
#> 62   actor9 actor19            1
#> 63   actor9 actor30            1
#> 64   actor9 actor56            1
#> 65  actor10 actor13            1
#> 66  actor10 actor17            1
#> 67  actor10 actor28            1
#> 68  actor10 actor31            1
#> 69  actor10 actor34            1
#> 70  actor10 actor46            1
#> 71  actor10 actor48            1
#> 72  actor10 actor52            1
#> 73  actor10 actor58            1
#> 74  actor13 actor16            1
#> 75  actor13 actor44            1
#> 76  actor13 actor56            1
#> 77  actor14 actor21            1
#> 78  actor14 actor22            1
#> 79  actor14 actor37            1
#> 80  actor14 actor52            1
#> 81  actor14 actor55            1
#> 82  actor16 actor56            1
#> 83  actor17 actor27            1
#> 84  actor17 actor48            1
#> 85  actor18 actor19            1
#> 86  actor19 actor29            1
#> 87  actor19 actor30            1
#> 88  actor19 actor31            1
#> 89  actor19 actor36            1
#> 90  actor19 actor44            1
#> 91  actor19 actor46            1
#> 92  actor20 actor48            1
#> 93  actor22 actor25            1
#> 94  actor22 actor50            1
#> 95  actor22 actor58            1
#> 96  actor24 actor56            1
#> 97  actor25 actor36            1
#> 98  actor25 actor46            1
#> 99  actor27 actor41            1
#> 100 actor28 actor31            1
#> 101 actor28 actor42            1
#> 102 actor28 actor56            1
#> 103 actor29 actor41            1
#> 104 actor31 actor56            1
#> 105 actor37 actor56            1
#> 106 actor44 actor51            1
#> 107 actor45 actor54            1
#> 108 actor52 actor58            1
#> 109  actor1 actor14            0
#> 110  actor1 actor77            0
#> 111  actor2 actor34            0
#> 112  actor2 actor57            0
#> 113  actor2 actor73            0
#> 114  actor3  actor7            0
#> 115  actor3 actor29            0
#> 116  actor3 actor36            0
#> 117  actor3 actor54            0
#> 118  actor3 actor60            0
#> 119  actor3 actor66            0
#> 120  actor4 actor51            0
#> 121  actor5 actor51            0
#> 122  actor6 actor29            0
#> 123  actor7 actor21            0
#> 124  actor7 actor23            0
#> 125  actor7 actor29            0
#> 126  actor7 actor60            0
#> 127  actor7 actor68            0
#> 128  actor7 actor72            0
#> 129  actor7 actor73            0
#> 130  actor7 actor78            0
#> 131  actor8 actor67            0
#> 132  actor8 actor73            0
#> 133  actor9 actor50            0
#> 134  actor9 actor57            0
#> 135  actor9 actor73            0
#> 136 actor10 actor38            0
#> 137 actor10 actor42            0
#> 138 actor10 actor57            0
#> 139 actor11 actor14            0
#> 140 actor11 actor61            0
#> 141 actor12 actor26            0
#> 142 actor12 actor28            0
#> 143 actor12 actor48            0
#> 144 actor13 actor19            0
#> 145 actor13 actor20            0
#> 146 actor13 actor33            0
#> 147 actor13 actor57            0
#> 148 actor13 actor69            0
#> 149 actor13 actor72            0
#> 150 actor13 actor73            0
#> 151 actor14 actor19            0
#> 152 actor14 actor25            0
#> 153 actor14 actor29            0
#> 154 actor14 actor36            0
#> 155 actor14 actor42            0
#> 156 actor14 actor45            0
#> 157 actor14 actor46            0
#> 158 actor14 actor58            0
#> 159 actor14 actor66            0
#> 160 actor15 actor60            0
#> 161 actor15 actor78            0
#> 162 actor16 actor17            0
#> 163 actor16 actor20            0
#> 164 actor16 actor28            0
#> 165 actor16 actor33            0
#> 166 actor16 actor42            0
#> 167 actor16 actor46            0
#> 168 actor16 actor48            0
#> 169 actor16 actor57            0
#> 170 actor16 actor66            0
#> 171 actor16 actor72            0
#> 172 actor16 actor73            0
#> 173 actor17 actor56            0
#> 174 actor17 actor70            0
#> 175 actor17 actor73            0
#> 176 actor18 actor57            0
#> 177 actor18 actor73            0
#> 178 actor19 actor42            0
#> 179 actor19 actor56            0
#> 180 actor19 actor77            0
#> 181 actor20 actor31            0
#> 182 actor20 actor33            0
#> 183 actor20 actor69            0
#> 184 actor20 actor73            0
#> 185 actor22 actor47            0
#> 186 actor22 actor49            0
#> 187 actor22 actor78            0
#> 188 actor23 actor27            0
#> 189 actor23 actor29            0
#> 190 actor23 actor41            0
#> 191 actor23 actor54            0
#> 192 actor24 actor33            0
#> 193 actor24 actor42            0
#> 194 actor24 actor57            0
#> 195 actor24 actor68            0
#> 196 actor24 actor72            0
#> 197 actor24 actor73            0
#> 198 actor24 actor74            0
#> 199 actor25 actor43            0
#> 200 actor26 actor31            0
#> 201 actor26 actor38            0
#> 202 actor27 actor29            0
#> 203 actor27 actor40            0
#> 204 actor27 actor60            0
#> 205 actor27 actor67            0
#> 206 actor27 actor68            0
#> 207 actor27 actor70            0
#> 208 actor28 actor48            0
#> 209 actor28 actor63            0
#> 210 actor28 actor66            0
#> 211 actor28 actor67            0
#> 212 actor28 actor68            0
#> 213 actor28 actor73            0
#> 214 actor29 actor60            0
#> 215 actor30 actor77            0
#> 216 actor31 actor38            0
#> 217 actor31 actor48            0
#> 218 actor33 actor57            0
#> 219 actor33 actor73            0
#> 220 actor34 actor72            0
#> 221 actor34 actor73            0
#> 222 actor34 actor77            0
#> 223 actor34 actor78            0
#> 224 actor35 actor58            0
#> 225 actor36 actor77            0
#> 226 actor37 actor68            0
#> 227 actor37 actor72            0
#> 228 actor37 actor73            0
#> 229 actor37 actor74            0
#> 230 actor37 actor75            0
#> 231 actor38 actor56            0
#> 232 actor38 actor72            0
#> 233 actor38 actor73            0
#> 234 actor40 actor41            0
#> 235 actor40 actor54            0
#> 236 actor41 actor60            0
#> 237 actor41 actor67            0
#> 238 actor41 actor68            0
#> 239 actor41 actor74            0
#> 240 actor41 actor75            0
#> 241 actor41 actor78            0
#> 242 actor42 actor49            0
#> 243 actor42 actor56            0
#> 244 actor42 actor57            0
#> 245 actor42 actor63            0
#> 246 actor42 actor65            0
#> 247 actor42 actor66            0
#> 248 actor42 actor68            0
#> 249 actor42 actor72            0
#> 250 actor42 actor74            0
#> 251 actor43 actor50            0
#> 252 actor44 actor56            0
#> 253 actor44 actor57            0
#> 254 actor44 actor73            0
#> 255 actor45 actor49            0
#> 256 actor45 actor50            0
#> 257 actor45 actor58            0
#> 258 actor46 actor52            0
#> 259 actor46 actor56            0
#> 260 actor46 actor57            0
#> 261 actor46 actor59            0
#> 262 actor46 actor73            0
#> 263 actor46 actor77            0
#> 264 actor48 actor56            0
#> 265 actor48 actor72            0
#> 266 actor49 actor58            0
#> 267 actor51 actor56            0
#> 268 actor51 actor57            0
#> 269 actor51 actor72            0
#> 270 actor51 actor73            0
#> 271 actor51 actor74            0
#> 272 actor54 actor60            0
#> 273 actor55 actor78            0
#> 274 actor56 actor57            0
#> 275 actor56 actor58            0
#> 276 actor56 actor60            0
#> 277 actor56 actor66            0
#> 278 actor56 actor72            0
#> 279 actor56 actor73            0
#> 280 actor56 actor74            0
#> 281 actor57 actor72            0
#> 282 actor57 actor73            0
#> 283 actor58 actor62            0
#> 284 actor58 actor68            0
#> 285 actor58 actor73            0
#> 286 actor58 actor77            0
#> 287 actor58 actor78            0
#> 288 actor58 actor79            0

# }