Helper function for plotting gaps and critical edges

Note that this only works for undirected graphs. Regardless of whether the input graph is directed it is treated as undirected graph.

plot_gaps_or_critical_dyads(
  net,
  edge_contribution,
  colour,
  title,
  lvl_attr = c("sesType"),
  cutoff = 2,
  subset_graph = "none",
  ...
)

Arguments

net
network object

edge_contribution
data frame providing edge contribution data

colour
colour code for the weighted edges

title
title of the plot

lvl_attr
nodal attribute specifying level information

cutoff
Cutoff point in contributions of an edge to the number of motifs above which to analyse gaps

subset_graph
Whether to subset the graph to only show nodes involved in gaps. One of "none" (no subset, default), "partial" (only focal level is subset) or "focal" (only focal level shown)

...
list of additional parameters to be passed to plotting function (see motifr::plot_mnet), e.g. label = TRUE

Value

A plot of gaps or critical edges, sized by weight in a multilevel network

See also