Simulate a random baseline

A specified number of random networks using a modified Erdős-Rényi model is computed. In each of the random networks motifs are counted. A dataframe with these counts is returned. Optionally, other models than the modified Erdős-Rényi model can be used.

simulate_baseline(
  net,
  motifs,
  n = 10,
  lvl_attr = "sesType",
  assume_sparse = TRUE,
  model = "erdos_renyi",
  level = -1,
  ergm_model = NULL,
  directed = NULL
)

Arguments

net
statnet network object

motifs
list of motif identifier strings

n
number of random graphs

lvl_attr
character string specifying the attribute name where level information is stored in statnet object.

assume_sparse
whether the random graphs shall be assumed to be sparse. used to find ideal counting function. defaults to TRUE.

model
baseline model to be used. Options are 'erdos_renyi', 'fixed_densities'. See vignette "random_baselines" for more details. Defaults to 'erdos_renyi'.

level
lvl_attr of the variable level for the Actor's Choice model

ergm_model
ergm model as for example fitted by calling ``ergm::ergm()``. Used when model is set to ergm to sample random networks.

directed
whether the graph shall be treated as a directed graph. Per default (NULL) this is determined automatically using network::is.directed.

Value

data frame with one column for each motif identifier string and one row for every computed random graph

Details

Note that when using the Actor's Choice model this function does not choose the variable level automatically. Use the ``level`` parameter to provide a valid level.

When using ERGM the parameter ``net`` is not used. Random networks are sampled in R using the ``ergm_model`` parameter.

Examples

simulate_baseline(ml_net, list("1,2[I.C]"), n = 10, directed = FALSE)

#>    1,2[I.C]
#> 1       176
#> 2       135
#> 3       168
#> 4       190
#> 5       200
#> 6       181
#> 7       211
#> 8       196
#> 9       181
#> 10      148