Simulate a random baseline

A specified number of random networks using a modified Erdős-Rényi model is computed. In each of the random networks motifs are counted. A dataframe with these counts is returned.

simulate_baseline(
  net,
  motifs,
  n = 10,
  lvl_attr = c("sesType"),
  assume_sparse = TRUE,
  model = "erdos_renyi"
)

Arguments

net
statnet network object

motifs
list of motif identifier strings

n
number of random graphs

lvl_attr
character vector specifying the attribute name where level information is stored in statnet object.

assume_sparse
whether the random graphs shall be assumed to be sparse. used to find ideal counting function

model
baseline model to be used

Value

data frame with one column for each motif identifier string and one row for every computed random graph

Examples

simulate_baseline(ml_net, list('1,2[I.C]'), n = 10)

#>    1,2[I.C]
#> 1       200
#> 2       163
#> 3       199
#> 4       162
#> 5       210
#> 6       168
#> 7       164
#> 8       137
#> 9       145
#> 10      234